Add awaits for calling services, move shortcuts to shortcut manager
