+++ plugin-flex-ts-template-v2/src/feature-library/keyboard-shortcuts/flex-hooks/strings/index.ts
@@ -54,6 +54,9 @@ export enum StringTemplates {
   CustomShortcutNavigateToKeyboardShortcuts = 'KeyboardMgrCustomShortcutNavigateToKeyboardShortcuts',
   CustomShortcutNavigateToTeamsView = 'KeyboardMgrCustomShortcutNavigateToTeamsView',
   CustomShortcutNavigateToQueuesView = 'KeyboardMgrCustomShortcutNavigateToQueuesView',
+  CustomShortcutToggleCallRecording = 'KeyboardMgrCustomShortcutToggleCallRecording',
+  CustomShortcutPauseCallRecording = 'KeyboardMgrCustomShortcutPauseCallRecording',
+  CustomShortcutResumeCallRecording = 'KeyboardMgrCustomShortcutResumeCallRecording',
 }
 
 export const stringHook = () => ({
@@ -112,6 +115,9 @@ export const stringHook = () => ({
     [StringTemplates.CustomShortcutNavigateToKeyboardShortcuts]: 'Navigate to keyboard shortcuts',
     [StringTemplates.CustomShortcutNavigateToTeamsView]: 'Navigate to Teams View',
     [StringTemplates.CustomShortcutNavigateToQueuesView]: 'Navigate to Real-time Queues View',
+    [StringTemplates.CustomShortcutToggleCallRecording]: 'Toggle call recording pause/resume',
+    [StringTemplates.CustomShortcutPauseCallRecording]: 'Pause call recording',
+    [StringTemplates.CustomShortcutResumeCallRecording]: 'Resume call recording',
   },
   'es-ES': esES,
   'es-MX': esMX,
