+++ plugin-flex-ts-template-v2/src/feature-library/pause-recording/flex-hooks/custom-action/pauseRecording.ts
@@ -16,6 +16,6 @@ export const registerPauseCallRecordingAction = async () => {
       return;
     }
 
-    pauseRecording(payload.task);
+    await pauseRecording(payload.task);
   });
 };
