+++ plugin-flex-ts-template-v2/src/feature-library/pause-recording/flex-hooks/custom-action/resumeRecording.ts
@@ -16,6 +16,6 @@ export const registerResumeCallRecordingAction = async () => {
       return;
     }
 
-    resumeRecording(payload.task);
+    await resumeRecording(payload.task);
   });
 };
